Make SourceParser class heritable + create and use JsSourceParser in runASTAnalysis
