+++ src/SourceParser.js
@@ -1,8 +1,5 @@
-// Import Third-party Dependencies
-import * as meriyah from "meriyah";
-
 // CONSTANTS
-const kParsingOptions = {
+export const kParsingOptions = {
   next: true,
   loc: true,
   raw: true,
@@ -42,50 +39,4 @@ export class SourceParser {
   #removeHTMLComment(str) {
     return str.replaceAll(/<!--[\s\S]*?(?:-->)/g, "");
   }
-
-  /**
-   * @param {object} options
-   * @param {boolean} options.isEcmaScriptModule
-   */
-  parseScript(options = {}) {
-    const {
-      isEcmaScriptModule
-    } = options;
-
-    try {
-      const { body } = meriyah.parseScript(
-        this.source,
-        {
-          ...kParsingOptions,
-          module: isEcmaScriptModule,
-          globalReturn: !isEcmaScriptModule
-        }
-      );
-
-      return body;
-    }
-    catch (error) {
-      const isIllegalReturn = error.description.includes("Illegal return statement");
-
-      if (error.name === "SyntaxError" && (
-        error.description.includes("The import keyword") ||
-        error.description.includes("The export keyword") ||
-        isIllegalReturn
-      )) {
-        const { body } = meriyah.parseScript(
-          this.source,
-          {
-            ...kParsingOptions,
-            module: true,
-            globalReturn: isIllegalReturn
-          }
-        );
-
-        return body;
-      }
-
-      throw error;
-    }
-  }
 }
-
